Finalizing

+++ nmd.py
@@ -10,8 +10,8 @@ def set_random_color():
 
 def main():
     letters = input("Enter your text: ")
-    intervalBetween = int(input("How long do you want each letter to show? "))
-    intervalAfter = int(input("How long do you want to wait after the last letter? "))
+    intervalBetween = float(input("How long do you want each letter to show? "))
+    intervalAfter = float(input("How long do you want to wait after the last letter? "))
 
     while True:
         for letter in letters:
